[nrf fromtree] net: mgmt: Convert the mgmt API to use 64-bit masks
Instead of using 32 bit enum values for event numbers, convert
the code to use 64 bit long bit fields. This means that the
user API is changed to use 64 bit event values instead of 32
bit event values.

(cherry picked from commit 5a9a39caf3bf5195a38194bb844a428eaa7f2bb8)

+++ include/zephyr/net/socket_net_mgmt.h
@@ -73,7 +73,7 @@ struct sockaddr_nm {
 	uintptr_t nm_pid;
 
 	/** net_mgmt mask */
-	uint32_t nm_mask;
+	uint64_t nm_mask;
 };
